+++ src/main/java/com/crudlogger/crudloggerstarter/aspect/logger/repository/LogEntityRepository.java
@@ -1,7 +1,7 @@
 package com.crudlogger.crudloggerstarter.aspect.logger.repository;
 
 import com.crudlogger.crudloggerstarter.aspect.logger.model.LogEntity;
-import com.crudlogger.crudloggerstarter.crud.repository.Long.AbstractRepositoryLong;
+import com.crudlogger.crudloggerstarter.crud.repository.biginit.AbstractRepositoryBigint;
 import org.springframework.stereotype.Repository;
 
 /**
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
  * @date 12.04.2023
  */
 @Repository
-public interface LogEntityRepository extends AbstractRepositoryLong<LogEntity> {
+public interface LogEntityRepository extends AbstractRepositoryBigint<LogEntity> {
 }
