+++ src/main/java/com/crudlogger/crudloggerstarter/crud/controller/biginit/AbstractControllerBigintImpl.java
@@ -1,4 +1,4 @@
-package com.crudlogger.crudloggerstarter.crud.controller.Abstract.String;
+package com.crudlogger.crudloggerstarter.crud.controller.biginit;
 
 import com.crudlogger.crudloggerstarter.aspect.ToLogger;
 import com.crudlogger.crudloggerstarter.aspect.logger.model.json.ActionDomainEnum;
@@ -6,10 +6,8 @@ import com.crudlogger.crudloggerstarter.aspect.logger.model.json.ActionEnum;
 import com.crudlogger.crudloggerstarter.aspect.logger.model.json.HttpMethodEnum;
 import com.crudlogger.crudloggerstarter.crud.dto.AbstractResponseDTO;
 import com.crudlogger.crudloggerstarter.crud.dto.specification.request.SearchRequest;
-import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityLong;
-import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityString;
-import com.crudlogger.crudloggerstarter.crud.service.Long.AbstractServiceLong;
-import com.crudlogger.crudloggerstarter.crud.service.string.AbstractServiceString;
+import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityBigint;
+import com.crudlogger.crudloggerstarter.crud.service.biginit.AbstractServiceBigint;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,10 +15,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @RequiredArgsConstructor
-public abstract class AbstractControllerStringImpl<E extends AbstractEntityString, S extends AbstractServiceString<E>>
-        implements AbstractControllerString<E> {
+public abstract class AbstractControllerBigintImpl<E extends AbstractEntityBigint, S extends AbstractServiceBigint<E>>
+        implements AbstractControllerBigint<E> {
 
-    protected final AbstractServiceString<E> service;
+    protected final AbstractServiceBigint<E> service;
 
     @Override
     @ToLogger(action = ActionEnum.READ, actionDomain = ActionDomainEnum.CRUD, httpMethod = HttpMethodEnum.GET)
@@ -30,7 +28,7 @@ public abstract class AbstractControllerStringImpl<E extends AbstractEntityStrin
 
     @Override
     @ToLogger(action = ActionEnum.READ, actionDomain = ActionDomainEnum.CRUD, httpMethod = HttpMethodEnum.GET)
-    public ResponseEntity<E> getOne(@PathVariable String id) {
+    public ResponseEntity<E> getOne(@PathVariable Long id) {
         return ResponseEntity.ok(service.findById(id));
     }
 
@@ -54,7 +52,7 @@ public abstract class AbstractControllerStringImpl<E extends AbstractEntityStrin
 
     @Override
     @ToLogger(action = ActionEnum.DELETE, actionDomain = ActionDomainEnum.CRUD, httpMethod = HttpMethodEnum.DELETE, returnResponse = true)
-    public ResponseEntity<E> delete(@PathVariable String id) {
+    public ResponseEntity<E> delete(@PathVariable Long id) {
         return ResponseEntity.ok(service.delete(id));
     }
 }
