+++ src/main/java/com/crudlogger/crudloggerstarter/crud/model/entity/AbstractEntityBigint.java
@@ -15,7 +15,7 @@ import java.io.Serializable;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public abstract class AbstractEntityLong implements Serializable {
+public abstract class AbstractEntityBigint implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     protected Long id;
