+++ src/main/java/com/crudlogger/crudloggerstarter/crud/repository/biginit/AbstractRepositoryBigint.java
@@ -1,6 +1,6 @@
-package com.crudlogger.crudloggerstarter.crud.repository.Long;
+package com.crudlogger.crudloggerstarter.crud.repository.biginit;
 
-import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityLong;
+import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityBigint;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.NoRepositoryBean;
@@ -10,6 +10,6 @@ import org.springframework.data.repository.NoRepositoryBean;
  * @date 29.04.2023
  */
 @NoRepositoryBean
-public interface AbstractRepositoryLong<E extends AbstractEntityLong> extends JpaRepository<E, Long>,
+public interface AbstractRepositoryBigint<E extends AbstractEntityBigint> extends JpaRepository<E, Long>,
         JpaSpecificationExecutor<E> {
 }
