+++ src/main/java/com/crudlogger/crudloggerstarter/crud/repository/integer/AbstractRepositoryInteger.java
@@ -1,7 +1,6 @@
 package com.crudlogger.crudloggerstarter.crud.repository.integer;
 
 import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityInteger;
-import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityLong;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.NoRepositoryBean;
