+++ src/main/java/com/crudlogger/crudloggerstarter/crud/service/integer/AbstractServiceInteger.java
@@ -3,7 +3,6 @@ package com.crudlogger.crudloggerstarter.crud.service.integer;
 import com.crudlogger.crudloggerstarter.crud.dto.AbstractResponseDTO;
 import com.crudlogger.crudloggerstarter.crud.dto.specification.request.SearchRequest;
 import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityInteger;
-import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityLong;
 
 public interface AbstractServiceInteger<E extends AbstractEntityInteger> {
     E save(E entity);
