+++ src/main/java/com/crudlogger/crudloggerstarter/crud/service/string/AbstractServiceStringImpl.java
@@ -4,9 +4,7 @@ import com.crudlogger.crudloggerstarter.crud.controller.advice.*;
 import com.crudlogger.crudloggerstarter.crud.dto.AbstractResponseDTO;
 import com.crudlogger.crudloggerstarter.crud.dto.specification.SearchSpecification;
 import com.crudlogger.crudloggerstarter.crud.dto.specification.request.SearchRequest;
-import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityLong;
 import com.crudlogger.crudloggerstarter.crud.model.entity.AbstractEntityString;
-import com.crudlogger.crudloggerstarter.crud.repository.Long.AbstractRepositoryLong;
 import com.crudlogger.crudloggerstarter.crud.repository.string.AbstractRepositoryString;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
